+++ src/async/distinct-async.ts
@@ -1,5 +1,5 @@
 import { resolverAsync } from 'augmentative-iterable';
-import { distinctRecipe, orderedDistinctRecipe } from '../recipes';
+import { distinctRecipe } from '../recipes';
 import { allAsync } from './all-async';
 import { filterAsync } from './filter-async';
 import { hasLessOrExactlyAsync } from './has-less-or-exactly-async';
@@ -7,17 +7,13 @@ import { mapAsync } from './map-async';
 import { partitionAsync } from './partition-async';
 
 function distinctAsyncRecipe(filterOrAll: Function) {
-  return distinctRecipe(
+  return distinctRecipe({
+    map: mapAsync,
+    resolver: resolverAsync,
+    partition: partitionAsync,
     filterOrAll,
-    resolverAsync,
-    orderedDistinctRecipe({
-      map: mapAsync,
-      resolver: resolverAsync,
-      partition: partitionAsync,
-      filterOrAll,
-      hasLessOrExactly: hasLessOrExactlyAsync,
-    }),
-  );
+    hasLessOrExactly: hasLessOrExactlyAsync,
+  });
 }
 
 export const distinctAsync = distinctAsyncRecipe(filterAsync);
