+++ src/recipes/distinct-recipe.ts
@@ -1,6 +1,8 @@
 import { identity } from '../utils';
 import { AnyIterable } from 'augmentative-iterable';
 import { AnyMapper, isOrderAssured, ResolverType } from '../types-internal';
+import { DistinctIngredients } from './ingredients';
+import { orderedOperationRecipe } from './ordered-operation-recipe';
 
 function inc(map: Map<any, number>, y: any) {
   const result = (map.get(y) || 0) + 1;
@@ -17,11 +19,30 @@ export function incPredicate<T>(
   return (x: T) => resolver(mapper(x), (y) => inc(map, y) <= maxOcurrences);
 }
 
-export function distinctRecipe(
-  filterOrAll: Function,
-  resolver: ResolverType,
-  ordered: Function,
-) {
+function orderedDistinctRecipe({
+  map,
+  resolver,
+  partition,
+  filterOrAll,
+  hasLessOrExactly,
+}: DistinctIngredients) {
+  const ordered = orderedOperationRecipe(map, resolver, partition);
+  return function distinct<T>(
+    this: AnyIterable<T>,
+    mapper: AnyMapper<any>,
+    maxOcurrences: number,
+  ) {
+    const partitioned = ordered.call(this, mapper);
+    return filterOrAll.call(partitioned, (part: AnyIterable<any>) =>
+      hasLessOrExactly.call(part, maxOcurrences),
+    );
+  };
+}
+
+export function distinctRecipe(ingredients: DistinctIngredients) {
+  const ordered = orderedDistinctRecipe(ingredients);
+  const { filterOrAll, resolver } = ingredients;
+
   return function distinct<T, R>(
     this: AnyIterable<T>,
     mapper: AnyMapper<T> | number = identity,
