+++ src/sync/distinct.ts
@@ -1,5 +1,5 @@
 import { resolver } from 'augmentative-iterable';
-import { distinctRecipe, orderedDistinctRecipe } from '../recipes';
+import { distinctRecipe } from '../recipes';
 import { all } from './all';
 import { filter } from './filter';
 import { hasLessOrExactly } from './has-less-or-exactly';
@@ -7,17 +7,13 @@ import { map } from './map';
 import { partition } from './partition';
 
 function distinctSyncRecipe(filterOrAll: Function) {
-  return distinctRecipe(
-    filterOrAll,
+  return distinctRecipe({
+    map,
     resolver,
-    orderedDistinctRecipe({
-      map,
-      resolver,
-      partition,
-      filterOrAll,
-      hasLessOrExactly,
-    }),
-  );
+    partition,
+    filterOrAll,
+    hasLessOrExactly,
+  });
 }
 
 export const distinct = distinctSyncRecipe(filter);
